refactor: extract data integrity hooks
